requirements.txt
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-cov==4.1.0